+++ crates/sncast/src/helpers/mod.rs
@@ -1,3 +1,2 @@
 pub mod constants;
-pub mod response_structs;
 pub mod scarb_utils;
