+++ crates/sncast/src/response/structs.rs
