+++ crates/sncast/src/starknet_commands/account/create.rs
@@ -6,8 +6,8 @@ use camino::Utf8PathBuf;
 use clap::Args;
 use serde_json::json;
 use sncast::helpers::constants::{CREATE_KEYSTORE_PASSWORD_ENV_VAR, OZ_CLASS_HASH};
-use sncast::helpers::response_structs::{AccountCreateResponse, Decimal, Hex};
 use sncast::helpers::scarb_utils::CastConfig;
+use sncast::response::structs::{AccountCreateResponse, Decimal, Hex};
 use sncast::{extract_or_generate_salt, get_chain_id, get_keystore_password, parse_number};
 use starknet::accounts::{AccountFactory, OpenZeppelinAccountFactory};
 use starknet::core::types::{FeeEstimate, FieldElement};
