+++ crates/sncast/src/starknet_commands/account/delete.rs
@@ -3,8 +3,8 @@ use camino::Utf8PathBuf;
 use clap::Args;
 use promptly::prompt;
 use serde_json::Map;
-use sncast::helpers::response_structs::AccountDeleteResponse;
 use sncast::helpers::scarb_utils::get_scarb_manifest;
+use sncast::response::structs::AccountDeleteResponse;
 use std::fs::File;
 use std::io::Read;
 use std::io::Write;
