+++ crates/sncast/src/starknet_commands/declare.rs
@@ -2,9 +2,10 @@ use anyhow::{anyhow, Context, Result};
 use camino::Utf8PathBuf;
 use clap::Args;
 use scarb_artifacts::{get_contracts_map, ScarbCommand};
-use sncast::helpers::response_structs::Hex;
 use sncast::helpers::scarb_utils::get_package_metadata;
-use sncast::helpers::{response_structs::DeclareResponse, scarb_utils::get_scarb_manifest};
+use sncast::helpers::scarb_utils::get_scarb_manifest;
+use sncast::response::structs::DeclareResponse;
+use sncast::response::structs::Hex;
 use sncast::{apply_optional, handle_rpc_error, handle_wait_for_tx, WaitForTx};
 use starknet::accounts::AccountError::Provider;
 use starknet::accounts::{ConnectedAccount, Declaration};
