+++ crates/sncast/src/starknet_commands/invoke.rs
@@ -1,7 +1,7 @@
 use anyhow::{anyhow, Result};
 use clap::Args;
 
-use sncast::helpers::response_structs::{Hex, InvokeResponse};
+use sncast::response::structs::{Hex, InvokeResponse};
 use sncast::{apply_optional, handle_rpc_error, handle_wait_for_tx, WaitForTx};
 use starknet::accounts::AccountError::Provider;
 use starknet::accounts::{Account, Call, ConnectedAccount, Execution, SingleOwnerAccount};
