+++ crates/sncast/src/starknet_commands/multicall/new.rs
@@ -1,8 +1,9 @@
 use anyhow::{bail, Result};
 use camino::Utf8PathBuf;
 use clap::Args;
-use sncast::helpers::constants::DEFAULT_MULTICALL_CONTENTS;
-use sncast::helpers::response_structs::MulticallNewResponse;
+use sncast::{
+    helpers::constants::DEFAULT_MULTICALL_CONTENTS, response::structs::MulticallNewResponse,
+};
 
 #[derive(Args, Debug)]
 #[command(about = "Generate a template for the multicall .toml file", long_about = None)]
